# api/openapi.yaml
openapi: 3.0.3
info:
  title: ASM Platform API
  version: 0.8.0
  description: |
    Contract for the HTTP routes currently mounted by cmd/api. Discovery scope,
    task template, and task run management are implemented below service level
    but are not listed here until HTTP handlers are added.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/users/{user_id}/projects:
    get:
      tags: [platform-users]
      summary: List a user's project memberships without modifying them.
      parameters: [{ $ref: "#/components/parameters/UserID" }]
      responses:
        "200": { description: Project membership list., content: { application/json: { schema: { $ref: "#/components/schemas/PlatformUserProjectsEnvelope" } } } }
        "400": { description: Invalid user id., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Missing or cross-tenant user., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
...
  /workspace/summary:
    get:
      tags: [workspace]
      summary: Return an actor-scoped cross-project workspace summary.
      description: Ordinary members are aggregated only across their memberships. Explicit cross-project administrators are still restricted to their tenant and the read is audited.
      responses:
        "200":
          description: Workspace summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSummaryEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects:
    get:
      tags: [projects]
      summary: List projects visible to the current principal.
      parameters:
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paged projects, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectPageEnvelope" } } } }
    post:
      tags: [projects]
      summary: Create a draft project and initialize its owner membership.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreateRequest" }
      responses:
        "201": { description: Draft project, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /projects/{project_id}:
    get:
      tags: [projects]
      summary: Get a project profile.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200": { description: Project, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectEnvelope" } } } }
    patch:
      tags: [projects]
      summary: Update editable project fields.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateRequest" }
      responses:
        "200": { description: Updated project, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectEnvelope" } } } }
  /projects/{project_id}/transitions:
    post:
      tags: [projects]
      summary: Transition project lifecycle state.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, suspended, archived] }
                reason: { type: string, maxLength: 512 }
      responses:
        "200": { description: Updated project, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { description: Project onboarding is incomplete, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /projects/{project_id}/onboarding-status:
    get:
      tags: [projects]
      summary: Return project activation readiness.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200": { description: Readiness, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
  /projects/{project_id}/capabilities:
    get:
      tags: [projects]
      summary: Return effective capabilities for the current actor in one project.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200":
          description: Effective project role, permissions and activation readiness.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCapabilitiesEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{project_id}/subjects:
    get:
      tags: [projects]
      summary: List project subjects.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200": { description: Subjects, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
    post:
      tags: [projects]
      summary: Create a project subject.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectSubjectRequest" } } } }
      responses:
        "201": { description: Subject, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
  /projects/{project_id}/subjects/{id}:
    patch:
      tags: [projects]
      summary: Update a project subject.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/ID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectSubjectRequest" } } } }
      responses:
        "200": { description: Subject, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
  /projects/{project_id}/domains:
    get:
      tags: [projects]
      summary: List registrable root-domain profiles.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200": { description: Domains, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
    post:
      tags: [projects]
      summary: Add a root-domain profile and idempotently materialize its domain asset.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectDomainRequest" } } } }
      responses:
        "201": { description: Domain profile, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
        "422": { description: Invalid registrable domain, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /projects/{project_id}/domains/{id}:
    patch:
      tags: [projects]
      summary: Update root-domain profile metadata.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/ID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectDomainUpdateRequest" } } } }
      responses:
        "200": { description: Domain profile, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
  /projects/{project_id}/icp-filings:
    get:
      tags: [projects]
      summary: List ICP filing profiles.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      responses:
        "200": { description: ICP filings, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
    post:
      tags: [projects]
      summary: Create an ICP filing and associate root-domain profiles.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ICPRequest" } } } }
      responses:
        "201": { description: ICP filing, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
  /projects/{project_id}/icp-filings/{id}:
    patch:
      tags: [projects]
      summary: Update an ICP filing and replace domain associations.
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/ID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ICPRequest" } } } }
      responses:
        "200": { description: ICP filing, content: { application/json: { schema: { $ref: "#/components/schemas/GenericEnvelope" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
...
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
    UserID:
      name: user_id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
    PageNumber:
      name: page_number
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Authentication failed or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The principal cannot access the requested tenant or project scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: State conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    GenericEnvelope:
      type: object
      required: [request_id, data]
      properties:
        request_id: { type: string }
        data: {}
    ErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    PlatformUserProject:
      type: object
      required: [id, project_code, name, role, status, updated_at]
      properties:
        id: { type: integer, format: uint64 }
        project_code: { type: string }
        name: { type: string }
        role: { type: string, enum: [project_owner, security_ops, developer, viewer] }
        status: { type: string }
        updated_at: { type: string, format: date-time }
    PlatformUserProjectsEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/PlatformUserProject" } }
...
    ProjectEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Project" }
    ProjectPageEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/Project" } }
                total: { type: integer }
                page_size: { type: integer }
                page_number: { type: integer }
    WorkspaceProject:
      type: object
      required: [id, project_code, name, owner_user_id, business_unit, criticality, status, updated_at]
      properties:
        id: { type: integer, format: uint64 }
        project_code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 255 }
        owner_user_id: { type: string, maxLength: 64 }
        business_unit: { type: string, maxLength: 128 }
        criticality: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [draft, active, suspended, archived] }
        updated_at: { type: string, format: date-time }
    WorkspaceSummaryEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [projects, assets, risks, tickets, recent_projects]
              properties:
                projects:
                  type: object
                  required: [total, active, draft, suspended]
                  properties:
                    total: { type: integer, format: int64, minimum: 0 }
                    active: { type: integer, format: int64, minimum: 0 }
                    draft: { type: integer, format: int64, minimum: 0 }
                    suspended: { type: integer, format: int64, minimum: 0 }
                assets:
                  type: object
                  required: [total]
                  properties:
                    total: { type: integer, format: int64, minimum: 0 }
                risks:
                  type: object
                  required: [open, critical_high, overdue]
                  properties:
                    open: { type: integer, format: int64, minimum: 0 }
                    critical_high: { type: integer, format: int64, minimum: 0 }
                    overdue: { type: integer, format: int64, minimum: 0 }
                tickets:
                  type: object
                  required: [open, overdue]
                  properties:
                    open: { type: integer, format: int64, minimum: 0 }
                    overdue: { type: integer, format: int64, minimum: 0 }
                recent_projects:
                  type: array
                  items: { $ref: "#/components/schemas/WorkspaceProject" }
    ProjectCapabilitiesEnvelope:
      allOf:
        - $ref: "#/components/schemas/GenericEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [role, permissions, can_activate, onboarding_missing]
              properties:
                role: { type: string }
                permissions:
                  type: array
                  uniqueItems: true
                  items: { type: string }
                can_activate: { type: boolean }
                onboarding_missing:
                  type: array
                  items: { type: string }
    ProjectCreateRequest:
      type: object
      required: [project_code, name]
      properties:
        project_code: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{0,63}$" }
        name: { type: string, maxLength: 255 }
        owner_user_id: { type: string, maxLength: 64 }
        business_unit: { type: string, maxLength: 128 }
        criticality: { type: string, enum: [low, medium, high, critical], default: medium }
        description: { type: string, maxLength: 65535 }
    ProjectUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        owner_user_id: { type: string, maxLength: 64 }
        business_unit: { type: string, maxLength: 128 }
        criticality: { type: string, enum: [low, medium, high, critical] }
        description: { type: string, maxLength: 65535 }
    ProjectSubjectRequest:
      type: object
      required: [subject_name]
      properties:
        subject_name: { type: string, maxLength: 255 }
        subject_type: { type: string, enum: [company, government, institution, individual, other] }
        registration_code: { type: string, maxLength: 64 }
        country_code: { type: string, minLength: 2, maxLength: 2 }
        region: { type: string, maxLength: 128 }
        is_primary: { type: boolean }
        verification_status: { type: string, enum: [unverified, verified, mismatch] }
        source: { type: string, maxLength: 64 }
        verified_at: { type: string, format: date-time, nullable: true }
        evidence_summary: { type: string, maxLength: 1024 }
    ProjectDomainRequest:
      type: object
      required: [domain]
      properties:
        domain: { type: string, maxLength: 253 }
        subject_id: { type: integer, format: uint64 }
        is_primary: { type: boolean }
        ownership_status: { type: string, enum: [unverified, verified, mismatch] }
        source: { type: string, maxLength: 64 }
        evidence_summary: { type: string, maxLength: 1024 }
    ProjectDomainUpdateRequest:
      type: object
      properties:
        subject_id: { type: integer, format: uint64 }
        is_primary: { type: boolean }
        ownership_status: { type: string, enum: [unverified, verified, mismatch] }
        source: { type: string, maxLength: 64 }
        evidence_summary: { type: string, maxLength: 1024 }
    ICPRequest:
      type: object
      required: [subject_id, filing_no]
      properties:
        subject_id: { type: integer, format: uint64 }
        filing_no: { type: string, maxLength: 128 }
        filing_type: { type: string, enum: [filing, license] }
        website_name: { type: string, maxLength: 255 }
        status: { type: string, enum: [unverified, valid, invalid, cancelled] }
        approved_at: { type: string, format: date-time, nullable: true }
        source: { type: string, maxLength: 64 }
        verified_at: { type: string, format: date-time, nullable: true }
        evidence_summary: { type: string, maxLength: 1024 }
        domain_profile_ids:
          type: array
          maxItems: 100
          items: { type: integer, format: uint64 }
...